Revert "Revert "Getting AI code into main repo""
This reverts commit 359c22cd4805ced398567d54c8c6e03a3f35efc5.

needed to be able to add zombie AI from prev commits back into repo

+++ src/view/Camera.cpp
@@ -3,7 +3,7 @@
 #include "../view/Window.h"
 #include "../player/Marine.h"
 
-Camera::Camera(float w, float h) {
+Camera::Camera(int w, int h) {
     viewport = {0, 0, 0, 0};
     setViewSize(w, h);
 }
@@ -13,13 +13,13 @@ Camera::~Camera() {
 }
 
 // Move camera to player
-void Camera::move(float x, float y) {
+void Camera::move(int x, int y) {
     viewport.x = ( x + MARINE_HEIGHT / 2 ) - viewport.w / 2;
     viewport.y = ( y + MARINE_WIDTH / 2 ) - viewport.h / 2;
 }
 
 // Set camera width and height
-void Camera::setViewSize(float w, float h) {
+void Camera::setViewSize(int w, int h) {
     viewport.w = w;
     viewport.h = h;
 }
