+++ src/view/Camera.h
@@ -5,20 +5,19 @@
 class Camera {
 public:
 
-    void setViewSize(float w, float h); // Set camera width and height
-    void move(float x, float  y); // moves the camera based on player
+    void setViewSize(int w, int h); // Set camera width and height
+    void move(int x, int  y); // moves the camera based on player
 
-    Camera(float w, float h);
+    Camera(int w, int h);
     ~Camera();
 
-    float getX(){return viewport.x;};
-    float getY(){return viewport.y;};
-    float getW(){return viewport.w;};
-    float getH(){return viewport.h;};
+    int getX(){return viewport.x;};
+    int getY(){return viewport.y;};
+    int getW(){return viewport.w;};
+    int getH(){return viewport.h;};
     SDL_Rect getViewport() {return viewport;};
 
 private:
-
     SDL_Rect viewport; // Camera area of view
 
 };
